# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dexe Investing Api",
        "title": "Dexe Investing Api",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/info/global-token-whitelist": {
            "get": {
                "description": "Get Global Token Whitelist",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Info"
                ],
                "summary": "Get Global Token Whitelist",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.S"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.GlobalTokenWhitelist"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.E"
                        }
                    }
                }
            }
        },
        "/nonce/{wallet}": {
            "get": {
                "description": "Get Sign Nonce",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Nonce"
                ],
                "summary": "Get Sign Nonce",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User wallet address",
                        "name": "wallet",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.S"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service_routes.GetNonceResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.E"
                        }
                    }
                }
            }
        },
        "/pool-transfers/deposits/{wallet}": {
            "get": {
                "description": "Get Deposits By Wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PoolTransfers"
                ],
                "summary": "Get Deposits By Wallet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wallet Address",
                        "name": "wallet",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.S"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.PoolTransfer"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.E"
                        }
                    }
                }
            }
        },
        "/pool-transfers/withdrawals/{wallet}": {
            "get": {
                "description": "Get Withdrawals By Wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PoolTransfers"
                ],
                "summary": "Get Withdrawals By Wallet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wallet Address",
                        "name": "wallet",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.S"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.PoolTransfer"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.E"
                        }
                    }
                }
            }
        },
        "/pools": {
            "get": {
                "description": "Get All Trader Pools",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pools"
                ],
                "summary": "Get All Trader Pools",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.S"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Pool"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.E"
                        }
                    }
                }
            }
        },
        "/pools/{wallet}": {
            "get": {
                "description": "Get Pools By Creator Wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pools"
                ],
                "summary": "Get Pools By Creator Wallet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pool creator wallet",
                        "name": "wallet",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.S"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Pool"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.E"
                        }
                    }
                }
            }
        },
        "/trader/{traderWallet}": {
            "get": {
                "description": "Get Trader Info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Trader"
                ],
                "summary": "Get Trader Info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trader wallet address",
                        "name": "traderWallet",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.S"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service_routes.TraderInfoResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.E"
                        }
                    }
                }
            }
        },
        "/trades/{traderPool}": {
            "get": {
                "description": "Get Trades By PoolAdr",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Trades"
                ],
                "summary": "Get Trades By PoolAdr",
                "parameters": [
                    {
                        "type": "string",
                        "description": "traderPool Address",
                        "name": "traderPool",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.S"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/service_routes.responseTrade"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.E"
                        }
                    }
                }
            }
        },
        "/user/signup": {
            "post": {
                "description": "User SignUp [SIGN]",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User SignUp [SIGN]",
                "parameters": [
                    {
                        "type": "string",
                        "default": "iQxX3slnRg",
                        "description": "An authorization header",
                        "name": "x-morph",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User nickname",
                        "name": "nickname",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User wallet",
                        "name": "wallet",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Body with image",
                        "name": "file",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.S"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.E"
                        }
                    }
                }
            }
        },
        "/user/{wallet}": {
            "get": {
                "description": "Get User Info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get User Info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User wallet address",
                        "name": "wallet",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.S"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.E"
                        }
                    }
                }
            }
        },
        "/user/{wallet}/avatar": {
            "put": {
                "description": "Update User Avatar [SIGN]",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update User Avatar [SIGN]",
                "parameters": [
                    {
                        "type": "string",
                        "default": "iQxX3slnRg",
                        "description": "An authorization header",
                        "name": "x-morph",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User wallet address",
                        "name": "wallet",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Body with image",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.S"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.E"
                        }
                    }
                }
            }
        },
        "/user/{wallet}/nickname": {
            "put": {
                "description": "Update User Nickname [SIGN]",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update User Nickname [SIGN]",
                "parameters": [
                    {
                        "type": "string",
                        "default": "iQxX3slnRg",
                        "description": "An authorization header",
                        "name": "x-morph",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User wallet address",
                        "name": "wallet",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "New user nickname",
                        "name": "nickname",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.S"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.E"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.GlobalTokenWhitelist": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "blockNumber": {
                    "description": "blockNumber bigint,",
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "tx": {
                    "description": "tx character varying(255),",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Pool": {
            "type": "object",
            "properties": {
                "basicTokenAdr": {
                    "type": "string"
                },
                "blockNumber": {
                    "description": "blockNumber bigint,",
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "creatorAdr": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "dexeCommissionDen": {
                    "type": "integer"
                },
                "dexeCommissionNum": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "investorCommissionDen": {
                    "type": "integer"
                },
                "investorCommissionNum": {
                    "type": "integer"
                },
                "investorRestricted": {
                    "type": "boolean"
                },
                "isActualOn": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "poolAdr": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                },
                "totalSupply": {
                    "type": "string"
                },
                "traderCommissionDen": {
                    "type": "integer"
                },
                "traderCommissionNum": {
                    "type": "integer"
                },
                "tx": {
                    "description": "tx character varying(255),",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.PoolTransfer": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "blockNumber": {
                    "description": "blockNumber bigint,",
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "poolAdr": {
                    "type": "string"
                },
                "tx": {
                    "description": "tx character varying(255),",
                    "type": "string"
                },
                "type": {
                    "description": "enum('deposit', 'withdraw')",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "wallet": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "wallet": {
                    "type": "string"
                }
            }
        },
        "response.E": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "context": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.S": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "datetime": {
                    "type": "string"
                }
            }
        },
        "service_routes.GetNonceResponse": {
            "type": "object",
            "properties": {
                "nonce": {
                    "type": "integer"
                }
            }
        },
        "service_routes.TraderInfoResponse": {
            "type": "object",
            "properties": {
                "copiers": {
                    "type": "number"
                },
                "fund": {
                    "type": "integer"
                }
            }
        },
        "service_routes.responseTrade": {
            "type": "object",
            "properties": {
                "amountIn": {
                    "type": "string"
                },
                "amountOutMin": {
                    "type": "string"
                },
                "blockNumber": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "deadline": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "path": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "traderPool": {
                    "type": "string"
                },
                "tx": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    }
}